requirements.txt
numpy
sounddevice
RPi.GPIO